+++ src/isign/model/__init__.py
@@ -1,12 +1,10 @@
 """Various data structures."""
-from .base import Base
 from .error import Error
-from .error_field import ErrorField
+from .field_error import FieldError
 from .response import Response
 
 
 __all__ = [
-    "Base",
     "Error",
     "ErrorField",
     "Response",
