--- src/isign/model/base.py
@@ -1,6 +1,0 @@
-from typing import Dict
-
-
-class Base:
-    def __init__(self, raw: Dict) -> None:
-        self.raw = raw
