+++ src/isign/model/response.py
@@ -1,7 +1,10 @@
-from .base import Base
+from typing import Dict
 
 
-class Response(Base):
+class Response:
+    def __init__(self, raw: Dict) -> None:
+        self.raw = raw
+
     @property
     def status(self) -> str:
         return str(self.raw["status"])
