--- tests/isign/model/test_model_base.py
@@ -1,7 +1,0 @@
-from isign.model.base import Base
-
-
-def test_base_constructor() -> None:
-    raw = {"status": "ok", "foo": "bar"}
-    response = Base(raw)
-    assert response.raw == raw
